dom/xhr/tests/test_worker_xhr2.html

<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<!DOCTYPE HTML>
<html>
<!--
Tests of DOM Worker Threads XHR(Bug 450452 )
-->
<head>
  <title>Test for DOM Worker Threads XHR (Bug 450452 )</title>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=450452">DOM Worker Threads XHR (Bug 450452)</a>
<p id="display"></p>
<div id="content" style="display: none">

</div>
<pre id="test">
<script class="testbody" type="text/javascript">

  var worker = new Worker("xhr2_worker.js");

  worker.onmessage = function(event) {
    is(event.data, "done", "Got correct result");
    SimpleTest.finish();
  }
  worker.postMessage("worker_testXHR.txt");

  SimpleTest.waitForExplicitFinish();

</script>
</pre>
</body>
</html>
